Changes in shell grammar, reorganized source files, implment CommandTable
Previously, the command loop to read the input was implemented in the grammar in shell.y itself. Adding more and more actions to shell.y was making it hard to read. New grammar exits yyparse() after detecting a complete command. The shell loop itself is implemented in main.c. CommandTable has been implemented as a stack. It tracks command ID and the command which was entered.